Shell check fixes (#196)
* Update install.sh

fixes proposed by ShellCheck

* Update version_check.sh

fixes proposed by ShellCheck

* Update system_info.sh

fixes proposed by ShellCheck

* Update node_menu.sh

fixes proposed by ShellCheck

* Update uninstall_6_sparrow.sh

fixes proposed by ShellCheck

* Update uninstall_7_bisq.sh

fixes proposed by ShellCheck

* Update uninstall_8_glances.sh

fixes proposed by ShellCheck

* Update 0_install_system.sh

fixes proposed by ShellCheck

* Update 1_install_bitcoind.sh

fixes proposed by ShellCheck

* Update 2_install_fulcrum.sh

fixes proposed by ShellCheck

* Update 3_install_mempool.sh

fixes proposed by ShellCheck

* Update 4_install_lnd.sh

fixes proposed by ShellCheck

* Update 5_install_thunderhub.sh

fixes proposed by ShellCheck

* Update 6_install_sparrow.sh

fixes proposed by ShellCheck

* Update 7_install_bisq.sh

fixes proposed by ShellCheck

* Update 8_install_glances.sh

fixes proposed by ShellCheck

* Update install_bitfeed.sh

fixes proposed by ShellCheck

* Update install_btc-rpc-explorer.sh

fixes proposed by ShellCheck

* Update install_ln-visualizer.sh

fixes proposed by ShellCheck

* Update install_node_status.sh

fixes proposed by ShellCheck

* Update install_sparrow_tails.sh

fixes proposed by ShellCheck

* Update upgrade_1_bitcoin.sh

fixes proposed by ShellCheck

* Update upgrade_2_fulcrum.sh

fixes proposed by ShellCheck

* Update upgrade_3_mempool.sh

fixes proposed by ShellCheck

* Update upgrade_4_lnd.sh

fixes proposed by ShellCheck

* Update upgrade_5_thunderhub.sh

fixes proposed by ShellCheck

* Update upgrade_6_sparrow.sh

fixes proposed by ShellCheck

* Update upgrade_bitfeed.sh

fixes proposed by ShellCheck

* Update upgrade_btc-rpc-explorer.sh

fixes proposed by ShellCheck

+++ other/install_sparrow_tails.sh
@@ -112,8 +112,7 @@ echo -e "${G}Done.${NC}"
 echo
 echo -e "${Y}Verify the release files...${NC}"
 # checksum
-sha256sum --ignore-missing --check --status sparrow-"${SPARROW_VERSION}"-manifest.txt 
-if [ "$?" -eq 0 ]; then
+if sha256sum --ignore-missing --check --status sparrow-"${SPARROW_VERSION}"-manifest.txt; then
   echo -e "${G}Verification of release checksum in checksum file: OK${NC}"
 else
   echo -e "${R}Verification of release checksum: Not OK${NC}"
@@ -123,8 +122,7 @@ fi
 wget -O pgp_keys.asc https://keybase.io/craigraw/pgp_keys.asc
 # import into gpg || true
 gpg --import -q pgp_keys.asc
-gpg --verify sparrow-"${SPARROW_VERSION}"-manifest.txt.asc
-if [ "$?" != 0 ]; then
+if ! gpg --verify sparrow-"${SPARROW_VERSION}"-manifest.txt.asc; then
   echo -e "${R}The signature(s) for the downloaded file are not good signature. Exiting now.${NC}"
   exit 1
 else
@@ -139,7 +137,7 @@ echo
 echo -e "${Y}Extract release, move content to the Sparrow app dir in the Persistent dir...${NC}"
 tar xfz sparrow-"${SPARROW_VERSION}"-x86_64.tar.gz
 # move files into Sparrow application dir
-rm -rf "${SPARROW_APP_DIR}"/*
+rm -rf "${SPARROW_APP_DIR:?}"/*
 mv "${DOWNLOAD_DIR}"/Sparrow/* "${SPARROW_APP_DIR}"
 rmdir "${DOWNLOAD_DIR}"/Sparrow/
 echo -e "${G}Done.${NC}"
